continued implementation nutrition analysis flow 4(data is received in the presentation layer)

+++ domain/src/main/java/com/example/domain/nutrition_analysis/model/NutritionAnalysisResponseDomainModel.kt
@@ -100,6 +100,6 @@ data class TotalNutrientsKCalDomain(
 
 data class NutrientDomain(
     val label: String,
-    val quantity: Double,
+    val quantity: Int,
     val unit: String,
 )
